[refactor] Clean project
