+++ src/components/featurecard/sass/featurecard.sass
@@ -1,3 +1,5 @@
+@import './../../../sass/variables.omega.library'
+
 .featurecard
     display: flex
     flex-direction: column
@@ -27,7 +29,7 @@
     &__title
         color: #464646
         color: rgb(73, 98, 114)
-        font-family: 'Righteous', cursive
+        font-family: $font-righteous
         font-size: 26px
         margin: 32px
         margin-bottom: 0
