+++ src/components/footer/sass/footer.sass
@@ -1,3 +1,5 @@
+@import './../../../sass/variables.omega.library'
+
 .footer
     border-top: 4px solid #2b2b2b
     background-color: #181818
@@ -153,7 +155,7 @@
 
         &__about-nw
             display: block
-            text-align: center
+            margin-bottom: 24px
 
         &__logo
             line-height: 64px
