+++ src/components/header/sass/header.sass
@@ -1,5 +1,7 @@
+@import './../../../sass/variables.omega.library'
+
 .logo
-    font-family: 'Righteous', cursive
+    font-family: $font-righteous
     font-size: 32px
     margin-top: -2px
     color: #c53431
