+++ src/components/releasecard/sass/releasecard.sass
@@ -1,4 +1,4 @@
-@import './../../../sass/_fonts.sass'
+@import './../../../sass/variables.omega.library'
 
 .card
     text-decoration: none
