+++ src/pages/sass/home.sass
@@ -1,3 +1,5 @@
+@import './../../sass/variables.omega.library'
+
 .project-description
     background-size: cover
     background-repeat: no-repeat
