+++ src/pages/sass/installer.sass
@@ -1,3 +1,5 @@
+@import './../../sass/variables.omega.library'
+
 .installer
     background-color: #E8E8E8
     width: 50%
