+++ src/pages/sass/notfound.sass
@@ -1,3 +1,5 @@
+@import './../../sass/variables.omega.library'
+
 .notfound
     background-color: #fafafa
     padding-top: 32px
