+++ src/pages/sass/policy.sass
@@ -1,3 +1,5 @@
+@import './../../sass/variables.omega.library'
+
 .legal
     margin: 32px 25%
     background-color: #fff
