+++ src/pages/sass/releases.sass
@@ -1,3 +1,5 @@
+@import './../../sass/variables.omega.library'
+
 .releases
     &__banner
         padding: 48px 15%
