+++ src/pages/sass/simulator.sass
@@ -1,0 +1,4 @@
+@import './../../sass/variables.omega.library'
+
+.simulator
+    margin: 24px 0
