+++ src/pages/sass/wiki.sass
@@ -1,3 +1,5 @@
+@import './../../sass/variables.omega.library'
+
 .wiki
     margin: 0 calc(15% - 32px)
