+++ src/pages/simulator/FullSimulator.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react'
-import Calculator from '../../components/Calculator'
+import Calculator from '../../components/calculator/Calculator'
 
 export default class Simulator extends Component {
     constructor(props) {
