+++ src/sass/_fonts.sass
@@ -11,6 +11,6 @@
 // Code
 @import url('https://fonts.googleapis.com/css2?family=Fira+Code&display=swap')
 
-$font-righteous: 'Righteous', cursive
-$font-fira-sans: 'Fira Sans', sans-serif
-$font-code: 'Fira Code', sans-serif
+.material-icons,
+.material-icons-round
+    font-size: 16px
