+++ src/sass/omega.library.sass
@@ -4,16 +4,7 @@
 @import '_fonts.sass'
 @import '_animations.sass'
 
-@import 'components/_footer.sass'
-@import 'components/_calculator.sass'
-
-@import 'pages/_index.sass'
-@import 'pages/_policy.sass'
-@import 'pages/_releases.sass'
-@import 'pages/_installer.sass'
-@import 'pages/_simulator.sass'
-@import 'pages/_notfound.sass'
-@import 'pages/_wiki.sass'
+@import 'variables.omega.library'
 
 body, .body
     display: flex
