--- src/sass/pages/_simulator.sass
@@ -1,2 +1,0 @@
-.simulator
-    margin: 24px 0
