+++ src/sass/variables.omega.library.sass
@@ -1,0 +1,4 @@
+// Fonts
+$font-righteous: 'Righteous', cursive
+$font-fira-sans: 'Fira Sans', sans-serif
+$font-code: 'Fira Code', sans-serif
